implementation of needed changes
